Refactor sidebar rendering and remove redundant CSS loading in main.py; add sidebar component
